requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
yfinance>=1.3.0
pandas==2.2.3
APScheduler==3.10.4
